[update] extracting types to own file
